Vulkan: keep UV clamping; promote vTexLimits and vUV to highp
Restore clamp_coord() on the nearest-neighbour path and fix the actual
cause of the Android texture corruption (#992): vTexLimits is declared
mediump, and the rhi disables clamping under a texture window by sending
UINT16_MAX limits. Mobile drivers that honour RelaxedPrecision with
16-bit integers fold 65535 to -1, so clamp(uv, 0, -1) pins every texel
to base - 1. Desktop drivers keep the value at 32 bits, which is why the
clamp is a no-op there and the clamp itself was never the problem.
Declaring the limits highp (and UVRange in the vertex stage) fixes
nearest, every filtered path and the HD-texture path together, and keeps
the anti-bleed guard that non-windowed 2D draws rely on at upscaled
internal resolutions. vUV is promoted alongside: fp16 loses the sub-texel
fraction above u = 128 and can mis-select texels under upscaling.

Prebuilt SPIR-V regenerated with the same glslc (byte-identical output on
the unchanged shaders); only the TEXTURED primitive/feedback variants
change.

+++ rhi/shaders_vulkan/primitive.vert
@@ -6,13 +6,14 @@ layout(location = 1) in vec4 Color;
 layout(location = 2) in mediump uvec4 Window;
 layout(location = 3) in mediump ivec3 Param;
 layout(location = 4) in ivec4 UV;
-layout(location = 5) in mediump uvec4 UVRange;
+layout(location = 5) in highp uvec4 UVRange;
 
-layout(location = 1) out mediump vec2 vUV;
+layout(location = 1) out highp vec2 vUV;
 layout(location = 2) flat out mediump ivec3 vParam;
 layout(location = 3) flat out mediump ivec2 vBaseUV;
 layout(location = 4) flat out mediump ivec4 vWindow;
-layout(location = 5) flat out mediump ivec4 vTexLimits;
+/* highp: see vram.h. UINT16_MAX limits do not survive a 16-bit int. */
+layout(location = 5) flat out highp ivec4 vTexLimits;
 #endif
 layout(location = 0) out mediump vec4 vColor;
 /* Depth-cue sidecar: far colour + blend factor. t == 0 (the no-cue
